added automatic interface detection support

+++ setup.sh
@@ -1,16 +1,33 @@
 #!/usr/bin/bash
 
+# Edit the SSID and PASSPHRASE to connect to the internet
+# Or pass it as arguments to the script
 
-INTERFACE="wlx08beac12c548"
+SSID=$1
+PASSPHRASE=$2
 
-echo 'Setting up wifi for interface ' $INTERFACE
+export INTERFACE=$(ifconfig | awk '$1 ~ /^wl/ {print $1}' | rev | cut -c 2- | rev)
+
+if [ -z "$INTERFACE" ]
+then
+	echo 'No valid interface found. Please insert a wifi dongle or manually assign the INTERFACE variable in the script'
+	exit
+fi
+
+echo 'Interface found' $INTERFACE
+
+
+# Don't edit anything after this
+
+echo 'Setting up wifi for interface' $INTERFACE
+echo 'ESSID' $SSID ' PASSPHRASE' $PASSPHRASE 
 
 # Network manager needs to be disabled first
 sudo systemctl stop NetworkManager
 sudo systemctl disable NetworkManager
 
 # create a wpa file with ssid and passphrase
-wpa_passphrase $1 $2 | sudo tee /etc/wpa_supplicant.conf
+wpa_passphrase $SSID $PASSPHRASE | sudo tee /etc/wpa_supplicant.conf
 
 # connect to the wifi via wpa_supplicant
 sudo wpa_supplicant -B -c /etc/wpa_supplicant.conf -i $INTERFACE
@@ -19,8 +36,8 @@ sudo wpa_supplicant -B -c /etc/wpa_supplicant.conf -i $INTERFACE
 sudo dhclient $INTERFACE
 
 # set up the service files
-sudo cp wpa_supplicant.service /etc/systemd/system/wpa_supplicant.service
-sudo cp dhclient.service /etc/systemd/system/dhclient.service
+envsubst < wpa_supplicant.service | sudo tee /etc/systemd/system/wpa_supplicant.service
+envsubst < dhclient.service | sudo tee /etc/systemd/system/dhclient.service
 
 # enable the service files at boot
 sudo systemctl enable dhclient.service
